nothing

+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: hnakayam <hnakayam@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 13:59:26 by hnakayam          #+#    #+#             */
-/*   Updated: 2024/12/18 12:22:59 by hnakayam         ###   ########.fr       */
+/*   Updated: 2024/12/18 18:32:27 by hnakayam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -153,7 +153,7 @@ int		is_ray_facing_up(float ray_angle);
 int		is_ray_facing_down(float ray_angle);
 int		is_ray_facing_right(float ray_angle);
 int		is_ray_facing_left(float ray_angle);
-double	calculate_distance_between_two_points(int a_x, int a_y,
+float	calculate_distance_between_two_points(int a_x, int a_y,
 			int b_x, int b_y);
 int		has_wall_at(t_vars *vars,
 			int next_horizontal_touch_x, int next_horizontal_touch_y);
